refactor workflow service for improved structure and scaleability of code
