+++ portality/bll/services/workflow/quick_fail.py
@@ -1,0 +1,9 @@
+from portality.bll.services.workflow.core import State
+
+
+class QuickFailAwaitingAssignment(State):
+    pass
+
+
+class QuickFailCriteriaCheck(State):
+    pass
