+++ portality/bll/services/workflow/triage.py
@@ -1,16 +1,16 @@
-from typing import Type, Union
+from typing import Union
 
-from portality import models, constants
+from portality import constants
 from portality.bll import DOAJ
 from portality.bll.exceptions import AuthoriseException
-from portality.models import Account, EditorGroup
 
-# Generic state definition values
-ANY = "*"
-UNASSIGNED = "-"
-ASSIGNED = "+"
+from portality.bll.services.workflow.core import WorkflowEvent, State, Claim, Assign, Unclaim, Reassign, Unassign, \
+    Fail, ApplicationEdit, WorkflowAction
+from portality.models import Account, EditorGroup, WorkflowControl
 
+########################################
 # Triage module state definition values
+
 MODULE_TRIAGE = "triage"
 MODULE_TRIAGE_STAGE_IN_PROGRESS = "in_progress"
 MODULE_TRIAGE_STAGE_MINIMAL_REVIEW = "minimal_review"
@@ -20,265 +20,8 @@ MODULE_TRIAGE_STAGES = [
 ]
 MODULE_TRIAGE_EG = "Triage"
 
-class WorkflowControlStateQuery:
-    def __init__(self,
-                 module:str=None,
-                 stage:str=None,
-                 editor_group:str=None,
-                 reviewer:str=None,
-                 size:int=100):
-        self._module = module
-        self._stage = stage
-        self._editor_group = editor_group
-        self._reviewer = reviewer
-        self._size = size
-
-    @property
-    def size(self):
-        return self._size
-
-    @size.setter
-    def size(self, size:int):
-        self._size = size
-
-    def query(self):
-        must = []
-        must_not = []
-
-        if self._module is not None:
-            if self._module != ANY:
-                if self._module == UNASSIGNED:
-                    must_not.append({"exists": {"field": "state.module"}})
-                elif self._module == ASSIGNED:
-                    must.append({"exists": {"field": "state.module"}})
-                else:
-                    must.append({"term": {"state.module.exact": self._module}})
-
-        if self._stage is not None:
-            if self._stage != ANY:
-                if self._stage == UNASSIGNED:
-                    must_not.append({"exists": {"field": "state.stage"}})
-                elif self._stage == ASSIGNED:
-                    must.append({"exists": {"field": "state.stage"}})
-                else:
-                    must.append({"term": {"state.stage.exact": self._stage}})
-
-        if self._editor_group is not None:
-            if self._editor_group != ANY:
-                if self._editor_group == UNASSIGNED:
-                    must_not.append({"exists": {"field": "state.editor_group"}})
-                elif self._editor_group == ASSIGNED:
-                    must.append({"exists": {"field": "state.editor_group"}})
-                else:
-                    must.append({"term": {"state.editor_group.exact": self._editor_group}})
-
-        if self._reviewer is not None:
-            if self._reviewer != ANY:
-                if self._reviewer == UNASSIGNED:
-                    must_not.append({"exists": {"field": "state.reviewer"}})
-                elif self._reviewer == ASSIGNED:
-                    must.append({"exists": {"field": "state.reviewer"}})
-                else:
-                    must.append({"term": {"state.reviewer.exact": self._reviewer}})
-
-        bool = {}
-        if len(must) > 0:
-            bool["must"] = must
-        if len(must_not) > 0:
-            bool["must_not"] = must_not
-        q = {"query": {"bool": bool}}
-
-        q["sort"] = {"created_date": {"order": "asc"}}
-
-        return q
-
-class WorkflowEvent:
-    def __init__(self, actor:Union[str, Account]):
-        self._actor = actor
-
-    @property
-    def actor(self) -> Account:
-        if isinstance(self._actor, str):
-            self._actor = Account.pull(self._actor)
-        return self._actor
-
-    @property
-    def actor_id(self) -> str:
-        if isinstance(self._actor, str):
-            return self._actor
-        return self._actor.id
-
-class WorkflowAction:
-    def __init__(self, actor: Union[str, Account]):
-        self._actor = actor
-
-    @property
-    def actor(self) -> Account:
-        if isinstance(self._actor, str):
-            self._actor = Account.pull(self._actor)
-        return self._actor
-
-    @property
-    def actor_id(self) -> str:
-        if isinstance(self._actor, str):
-            return self._actor
-        return self._actor.id
-
-class State:
-    module = None
-    stage = None
-    editor_group = None
-    reviewer = None
-
-    events = []
-    actions = []
-
-    def __init__(self, wf_control:models.WorkflowControl, application:models.Application=None):
-        self._wf_control = wf_control
-        self._application = application
-
-    @classmethod
-    def query(cls):
-        return WorkflowControlStateQuery(cls.module, cls.stage, cls.editor_group, cls.reviewer)
-
-    @classmethod
-    def enter(cls, actor:Union[str, Account],
-              wf_control:models.WorkflowControl,
-              application:models.Application=None,
-              from_state: "State"=None):
-        instance = cls(wf_control, application)
-        instance.apply()
-        instance.audit(actor, from_state)
-        return instance
-
-    @classmethod
-    def matches(cls, wf_control:models.WorkflowControl):
-        # Check the module.
-        # If the wfc has no module, and the state definition requires a specific module, then it doesn't match.
-        # If the wfc has a module, and the state definition requires a specifc module, and it isn't this one, then it doesn't match
-        if wf_control.module is None:
-            if not (cls.module == ANY):
-                return False
-        else:
-            if not (cls.module == ANY or cls.module == wf_control.module):
-                return False
-
-        # if we get to here, module matches
-
-        # check the stage
-        # If the wfc has no stage, and the state definition requires a specific stage, then it doesn't match.
-        # If the wfc has a stage, and the state definition requires a specifc stage, and it isn't this one, then it doesn't match
-        if wf_control.stage is None:
-            if not (cls.stage == ANY):
-                return False
-        else:
-            if not (cls.stage == ANY or cls.stage == wf_control.stage):
-                return False
-
-        # by here, module and stage match
-
-        # check the editor group
-        # if the wfc has no eg, then if the allowed editor group is not ANY or UNASSIGNED), no match
-        # if the wfc has an eg, then if the allowed editor group is not ANY or ASSIGNED or the same as the wfc's eg, no match
-        if wf_control.editor_group is None:
-            if not (cls.editor_group == ANY or cls.editor_group == UNASSIGNED):
-                return False
-        else:
-            if not (cls.editor_group == ANY or cls.editor_group == ASSIGNED or cls.editor_group == wf_control.editor_group):
-                return False
-
-        # by here, module, stage and editor group match
-
-        if wf_control.reviewer is None:
-            if not (cls.reviewer == ANY or cls.reviewer == UNASSIGNED):
-                return False
-        else:
-            if not (cls.reviewer == ANY or cls.reviewer == ASSIGNED or cls.reviewer == wf_control.editor_group):
-                return False
-
-        return True
-
-    @property
-    def workflow_control(self):
-        return self._wf_control
-
-    @property
-    def application(self):
-        if self._application is None:
-            app_id = self.workflow_control.application_id
-            self._application = models.Application.pull(app_id)
-        return self._application
-
-    def apply(self):
-        pass
-
-    def exit(self, event:WorkflowEvent):
-        return self
-
-    def do(self, action:WorkflowAction):
-        return self
-
-    def get_audit_partial(self):
-        return {
-            "module": self.module,
-            "stage": self.stage,
-            "editor_group": self.workflow_control.editor_group,
-            "reviewer": self.workflow_control.reviewer
-        }
-
-    def audit(self, actor, from_state:"State"=None, date=None):
-        origin_data = None
-        if from_state is not None:
-            origin_data = from_state.get_audit_partial()
-        target_data = self.get_audit_partial()
-        self.workflow_control.add_audit(actor, target_data, origin_data, date)
-
-    def saveall(self, *args, **kwargs):
-        self.workflow_control.save(*args, **kwargs)
-        if self._application:
-            self._application.save(*args, **kwargs)
-
-
-class ReviewerAssignment(WorkflowEvent):
-    def __init__(self, actor:Union[str, Account], reviewer:Union[str, Account]):
-        super().__init__(actor)
-        self._reviewer = reviewer
-
-    @property
-    def reviewer(self) -> Account:
-        if isinstance(self._reviewer, str):
-            self._reviewer = Account.pull(self._reviewer)
-        return self._reviewer
-
-    @property
-    def reviewer_id(self) -> str:
-        if isinstance(self._reviewer, str):
-            return self._reviewer
-        return self._reviewer.id
-
-class Claim(WorkflowEvent): pass
-
-class Unclaim(WorkflowEvent): pass
-
-class Assign(ReviewerAssignment): pass
-
-class Reassign(ReviewerAssignment): pass
-
-class Unassign(WorkflowEvent): pass
-
-class Fail(WorkflowEvent):
-    def __init__(self, actor:Union[str, Account], note:str=None, embargo_end:str=None):
-        super().__init__(actor)
-        self._note = note
-        self._embargo_end = embargo_end
-
-    @property
-    def note(self):
-        return self._note
-
-    @property
-    def embargo_end(self):
-        return self._embargo_end
+######################################
+## Workflow events specific to Triage
 
 class MinimalReview(WorkflowEvent): pass
 
@@ -300,14 +43,14 @@ class Triaged(WorkflowEvent):
     def maned(self):
         return self._maned
 
-
-class ApplicationEdit(WorkflowAction): pass
+######################################
+## Triage Workflow States
 
 class AwaitingTriage(State):
     module = MODULE_TRIAGE
-    stage = ANY
+    stage = WorkflowControl.ANY
     editor_group = MODULE_TRIAGE_EG
-    reviewer = UNASSIGNED
+    reviewer = WorkflowControl.UNASSIGNED
 
     events = [Claim, Assign]
 
@@ -371,11 +114,12 @@ class AwaitingTriage(State):
         else:
             return TriageAssessmentInProgress.enter(event.actor, self._wf_control, self._application, self)
 
+
 class TriageAssessmentInProgress(State):
     module = MODULE_TRIAGE
     stage = MODULE_TRIAGE_STAGE_IN_PROGRESS
     editor_group = MODULE_TRIAGE_EG
-    reviewer = ASSIGNED
+    reviewer = WorkflowControl.ASSIGNED
 
     events = [Unclaim, Reassign, Unassign, Fail, MinimalReview]
     actions = [ApplicationEdit]
@@ -463,10 +207,10 @@ class TriageAssessmentInProgress(State):
         del wfc.reviewer
 
         # TODO: not clear what to do with application embargoes, perhaps these are a new type?
-
+        from portality.bll.services.workflow.rejected import Rejected
         return Rejected.enter(event.actor, self._wf_control, self._application, self)
 
-    def event_minimal_review(self, event:MinimalReview):
+    def event_minimal_review(self, event: MinimalReview):
         wfc = self.workflow_control
         if wfc.reviewer != event.actor_id:
             raise AuthoriseException(reason=AuthoriseException.NOT_AUTHORISED)
@@ -485,11 +229,12 @@ class TriageAssessmentInProgress(State):
         else:
             return self
 
+
 class TriageAssessmentMinimalReview(State):
     module = MODULE_TRIAGE
     stage = MODULE_TRIAGE_STAGE_MINIMAL_REVIEW
     editor_group = MODULE_TRIAGE_EG
-    reviewer = ASSIGNED
+    reviewer = WorkflowControl.ASSIGNED
 
     events = [Triaged, Unclaim, Reassign, Unassign, Fail, RescindMinimalReview]
     actions = [ApplicationEdit]
@@ -543,7 +288,7 @@ class TriageAssessmentMinimalReview(State):
         else:
             raise ValueError(f"Unknown action '{action}' for state '{type(self).__name__}'")
 
-    def event_triaged(self, event:Triaged):
+    def event_triaged(self, event: Triaged):
         wfc = self.workflow_control
         if wfc.reviewer != event.actor_id:
             raise AuthoriseException(reason=AuthoriseException.NOT_AUTHORISED)
@@ -565,6 +310,7 @@ class TriageAssessmentMinimalReview(State):
             wfc.editor_group = vessel.name
             wfc.reviewer = event.maned
 
+            from portality.bll.services.workflow.quick_fail import QuickFailCriteriaCheck
             return QuickFailCriteriaCheck.enter(event.actor, self._wf_control, self._application, self)
 
         elif event.editor_group is not None:
@@ -575,6 +321,7 @@ class TriageAssessmentMinimalReview(State):
             wfc.editor_group = eg.name
             del wfc.reviewer
 
+            from portality.bll.services.workflow.quick_fail import QuickFailAwaitingAssignment
             return QuickFailAwaitingAssignment.enter(event.actor, self._wf_control, self._application, self)
 
         raise ValueError("Triaged event must have either a target maned or a target editor group")
@@ -609,13 +356,14 @@ class TriageAssessmentMinimalReview(State):
             raise AuthoriseException(reason=AuthoriseException.NOT_AUTHORISED)
 
         appSvc = DOAJ.applicationService()
-        appSvc.reject_application(self.application, event.actor, event.note)
+        appSvc.reject_application(self.application, event.actor, note=event.note)
 
         del wfc.editor_group
         del wfc.reviewer
 
         # TODO: not clear what to do with application embargoes, perhaps these are a new type?
 
+        from portality.bll.services.workflow.rejected import Rejected
         return Rejected.enter(event.actor, self._wf_control, self._application, self)
 
     def event_rescind_minimal_review(self, event: RescindMinimalReview):
@@ -636,112 +384,3 @@ class TriageAssessmentMinimalReview(State):
             return self
         else:
             return self.event_rescind_minimal_review(RescindMinimalReview(action.actor_id))
-
-
-class QuickFailAwaitingAssignment(State):
-    pass
-
-class QuickFailCriteriaCheck(State):
-    pass
-
-MODULE_REJECTED = "rejected"
-MODULE_REJECTED_STAGE_REJECTED = "rejected"
-
-class Rejected(State):
-    module = MODULE_REJECTED
-    stage = MODULE_REJECTED_STAGE_REJECTED
-    editor_group = ANY
-    reviewer = ANY
-
-    def apply(self):
-        # Ensure the workflow control object is in the right state
-        wf_control = self.workflow_control
-        if wf_control.module != self.module:
-            wf_control.module = self.module
-
-        if wf_control.stage != self.stage:
-            wf_control.stage = self.stage
-
-        # back-compatibility for the original workflow
-        application = self.application
-
-        if application.application_status != constants.APPLICATION_STATUS_REJECTED:
-            appSvc = DOAJ.applicationService()
-            acc = None
-            if wf_control.reviewer is not None:
-                acc = models.Account.pull(wf_control.reviewer)
-            if acc is None:
-                # FIXME: what's the way to do this?
-                acc = models.Account.pull("system")
-            appSvc.reject_application(application, acc)
-
-        if wf_control.editor_group is not None:
-            if application.editor_group != wf_control.editor_group:
-                application.set_editor_group(wf_control.editor_group)
-        else:
-            application.remove_editor_group()
-
-        if wf_control.reviewer is not None:
-            if application.editor != wf_control.reviewer:
-                application.set_editor(wf_control.reviewer)
-        else:
-            application.remove_editor()
-
-
-class WorkflowService:
-    STATES = [
-        # Triage States
-        AwaitingTriage,
-        TriageAssessmentInProgress,
-        TriageAssessmentMinimalReview,
-
-        # Quick Fail States
-        QuickFailAwaitingAssignment,
-        QuickFailCriteriaCheck
-    ]
-
-    def iterate_state(self, state:Type[State]):
-        query = state.query()
-        for wfc in models.WorkflowControl.iterate_unstable(q=query.query()):
-            yield state(wfc)
-
-    def first_n_in_state(self, state:Type[State], n:int) -> list[State]:
-        query = state.query()
-        query.size = n
-        return [state(x) for x in models.WorkflowControl.object_query(q=query.query())]
-
-    def apply_event(self, wfc_id:str, event:WorkflowEvent, save=True):
-        state_instance = self.state_for_workflow_control(wfc_id)
-        if state_instance is None:
-            raise ValueError(f"No state found for workflow control with id '{wfc_id}'")
-
-        new_state = self.event(state_instance, event)
-        if save:
-            new_state.saveall()
-        return new_state
-
-    def state_for_workflow_control(self, wfc_id:str) -> Union[State, None]:
-        wfc = models.WorkflowControl.pull(wfc_id)
-        if wfc is None:
-            return None
-
-        for state in self.STATES:
-            if state.matches(wfc):
-                return state(wfc)
-
-        return None
-
-    def event(self, state_instance:State, event:WorkflowEvent) -> State:
-        if event.__class__ not in state_instance.events:
-            raise ValueError(f"Invalid event '{event}' for state '{type(state_instance).__name__}'")
-
-        new_state = state_instance.exit(event)
-        return new_state
-
-    def initialise_workflow(self, actor:Union[str, Account], application:models.Application) -> State:
-        wfc = models.WorkflowControl()
-        wfc.application_id = application.id
-        wfc.original_application = application
-        wfc.application_title = application.bibjson().title
-        initial_state = AwaitingTriage.enter(actor, wfc, application)
-        return initial_state
